Declare DuplicatedVotesPool as subclass of MessageLog
